% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateBootstrap.R
\name{ateBootstrap}
\alias{ateBootstrap}
\title{Compute bootstrap confidence intervals on the ATE parameter}
\description{
Output both the point estimate and the bootstrap confidence interval
}
\section{Super class}{
\code{\link[TMLEbootstrap:generalBootstrap]{TMLEbootstrap::generalBootstrap}} -> \code{ateBootstrap}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ateBootstrap$new()}}
\item \href{#method-bootstrap_once}{\code{ateBootstrap$bootstrap_once()}}
\item \href{#method-run_bootstrap}{\code{ateBootstrap$run_bootstrap()}}
\item \href{#method-clone}{\code{ateBootstrap$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="all_boot_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-all_boot_CI}{\code{TMLEbootstrap::generalBootstrap$all_boot_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="bootstrap">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-bootstrap}{\code{TMLEbootstrap::generalBootstrap$bootstrap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="center_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-center_CI}{\code{TMLEbootstrap::generalBootstrap$center_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="exact_bootstrap">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-exact_bootstrap}{\code{TMLEbootstrap::generalBootstrap$exact_bootstrap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="exact_bootstrap_paper">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-exact_bootstrap_paper}{\code{TMLEbootstrap::generalBootstrap$exact_bootstrap_paper()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="penalized_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-penalized_CI}{\code{TMLEbootstrap::generalBootstrap$penalized_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="penalized_CI_half">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-penalized_CI_half}{\code{TMLEbootstrap::generalBootstrap$penalized_CI_half()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="scale_adjust_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-scale_adjust_CI}{\code{TMLEbootstrap::generalBootstrap$scale_adjust_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="shift2">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-shift2}{\code{TMLEbootstrap::generalBootstrap$shift2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="sigma_mse">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-sigma_mse}{\code{TMLEbootstrap::generalBootstrap$sigma_mse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="spread">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-spread}{\code{TMLEbootstrap::generalBootstrap$spread()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateBootstrap$new(
  data,
  family_y,
  lambda1 = NULL,
  lambda2 = NULL,
  M1 = NULL,
  M2 = NULL,
  targeting = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bootstrap_once"></a>}}
\if{latex}{\out{\hypertarget{method-bootstrap_once}{}}}
\subsection{Method \code{bootstrap_once()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateBootstrap$bootstrap_once(self, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-run_bootstrap}{}}}
\subsection{Method \code{run_bootstrap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateBootstrap$run_bootstrap(n_bootstrap = 200, alpha = 0.05, kind = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateBootstrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
