% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ateTMLE.R
\name{ateTMLE}
\alias{ateTMLE}
\title{Compute bootstrap confidence intervals on the ATE parameter}
\description{
Compute bootstrap confidence intervals on the ATE parameter

Compute bootstrap confidence intervals on the ATE parameter
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ateTMLE$new()}}
\item \href{#method-initial_fit}{\code{ateTMLE$initial_fit()}}
\item \href{#method-initial_fit_pen_likeli}{\code{ateTMLE$initial_fit_pen_likeli()}}
\item \href{#method-plot_Q1W}{\code{ateTMLE$plot_Q1W()}}
\item \href{#method-target}{\code{ateTMLE$target()}}
\item \href{#method-compute_EIC}{\code{ateTMLE$compute_EIC()}}
\item \href{#method-inference_without_target}{\code{ateTMLE$inference_without_target()}}
\item \href{#method-compute_min_phi_ratio}{\code{ateTMLE$compute_min_phi_ratio()}}
\item \href{#method-clone}{\code{ateTMLE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$new(data, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial_fit"></a>}}
\if{latex}{\out{\hypertarget{method-initial_fit}{}}}
\subsection{Method \code{initial_fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$initial_fit(
  lambda1 = NULL,
  lambda2 = NULL,
  M1 = NULL,
  M2 = NULL,
  family_y,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial_fit_pen_likeli"></a>}}
\if{latex}{\out{\hypertarget{method-initial_fit_pen_likeli}{}}}
\subsection{Method \code{initial_fit_pen_likeli()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$initial_fit_pen_likeli(
  lambda1 = NULL,
  lambda2 = NULL,
  family_y,
  lambda_min_ratio = NULL,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_Q1W"></a>}}
\if{latex}{\out{\hypertarget{method-plot_Q1W}{}}}
\subsection{Method \code{plot_Q1W()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$plot_Q1W(foo = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target"></a>}}
\if{latex}{\out{\hypertarget{method-target}{}}}
\subsection{Method \code{target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_EIC"></a>}}
\if{latex}{\out{\hypertarget{method-compute_EIC}{}}}
\subsection{Method \code{compute_EIC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$compute_EIC(A, gk, Y, Qk, Q1k, Q0k, psi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inference_without_target"></a>}}
\if{latex}{\out{\hypertarget{method-inference_without_target}{}}}
\subsection{Method \code{inference_without_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$inference_without_target(
  data,
  Q_fit = NULL,
  g_fit = NULL,
  to_return = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_min_phi_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-compute_min_phi_ratio}{}}}
\subsection{Method \code{compute_min_phi_ratio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$compute_min_phi_ratio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
