% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneHyperparam.R
\name{ateTuneHyperparam}
\alias{ateTuneHyperparam}
\title{Use plateau method to choose the L1 penalty of HAL (ATE)}
\description{
Use plateau method to choose the L1 penalty of HAL (ATE)

Use plateau method to choose the L1 penalty of HAL (ATE)
}
\section{Super class}{
\code{\link[TMLEbootstrap:tuneHyperparam]{TMLEbootstrap::tuneHyperparam}} -> \code{AteTuneHyperparam}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ateTuneHyperparam$new()}}
\item \href{#method-add_lambda}{\code{ateTuneHyperparam$add_lambda()}}
\item \href{#method-clone}{\code{ateTuneHyperparam$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="get_lambda">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-get_lambda}{\code{TMLEbootstrap::tuneHyperparam$get_lambda()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="get_lambda_df">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-get_lambda_df}{\code{TMLEbootstrap::tuneHyperparam$get_lambda_df()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="get_Psi_df">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-get_Psi_df}{\code{TMLEbootstrap::tuneHyperparam$get_Psi_df()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="get_value">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-get_value}{\code{TMLEbootstrap::tuneHyperparam$get_value()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="plot_CI">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-plot_CI}{\code{TMLEbootstrap::tuneHyperparam$plot_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="plot_width">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-plot_width}{\code{TMLEbootstrap::tuneHyperparam$plot_width()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="select_lambda_pleateau_wald">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-select_lambda_pleateau_wald}{\code{TMLEbootstrap::tuneHyperparam$select_lambda_pleateau_wald()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="tuneHyperparam" data-id="select_lambda_psi_grad">}\href{../../TMLEbootstrap/html/tuneHyperparam.html#method-select_lambda_psi_grad}{\code{TMLEbootstrap::tuneHyperparam$select_lambda_psi_grad()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTuneHyperparam$new(..., family_y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-add_lambda}{}}}
\subsection{Method \code{add_lambda()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTuneHyperparam$add_lambda(lambda_grid = NULL, to_parallel = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ateTuneHyperparam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
