% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDensityBootstrap.R
\name{avgDensityBootstrap}
\alias{avgDensityBootstrap}
\title{Compute bootstrap confidence intervals on the average squared density parameter}
\description{
Compute bootstrap confidence intervals on the average squared density parameter

Compute bootstrap confidence intervals on the average squared density parameter
}
\details{
Output both the point estimate and the bootstrap confidence interval
}
\section{Super class}{
\code{\link[TMLEbootstrap:generalBootstrap]{TMLEbootstrap::generalBootstrap}} -> \code{avgDensityBootstrap}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{avgDensityBootstrap$new()}}
\item \href{#method-bootstrap_once}{\code{avgDensityBootstrap$bootstrap_once()}}
\item \href{#method-run_bootstrap}{\code{avgDensityBootstrap$run_bootstrap()}}
\item \href{#method-clone}{\code{avgDensityBootstrap$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="all_boot_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-all_boot_CI}{\code{TMLEbootstrap::generalBootstrap$all_boot_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="bootstrap">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-bootstrap}{\code{TMLEbootstrap::generalBootstrap$bootstrap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="center_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-center_CI}{\code{TMLEbootstrap::generalBootstrap$center_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="exact_bootstrap">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-exact_bootstrap}{\code{TMLEbootstrap::generalBootstrap$exact_bootstrap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="exact_bootstrap_paper">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-exact_bootstrap_paper}{\code{TMLEbootstrap::generalBootstrap$exact_bootstrap_paper()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="penalized_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-penalized_CI}{\code{TMLEbootstrap::generalBootstrap$penalized_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="penalized_CI_half">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-penalized_CI_half}{\code{TMLEbootstrap::generalBootstrap$penalized_CI_half()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="scale_adjust_CI">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-scale_adjust_CI}{\code{TMLEbootstrap::generalBootstrap$scale_adjust_CI()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="shift2">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-shift2}{\code{TMLEbootstrap::generalBootstrap$shift2()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="sigma_mse">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-sigma_mse}{\code{TMLEbootstrap::generalBootstrap$sigma_mse()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="generalBootstrap" data-id="spread">}\href{../../TMLEbootstrap/html/generalBootstrap.html#method-spread}{\code{TMLEbootstrap::generalBootstrap$spread()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
bootstrap average density parameter;
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityBootstrap$new(
  x,
  epsilon_step = NULL,
  bin_width = 0.3,
  lambda_grid = NULL,
  M = NULL,
  targeting = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{vector of random samples.}

\item{\code{epsilon_step}}{step size of the TMLE (one-step)}

\item{\code{bin_width}}{binning width for the density super learner}

\item{\code{lambda_grid}}{pre-specified grid of L1 penalization for the HAL}

\item{\code{M}}{alternative to "lambda_grid". Specify the max variation norm of the HAL}

\item{\code{targeting}}{bool. whether to perform targeting on the density estimator. If FALSE, the density super learner estimate will be returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bootstrap_once"></a>}}
\if{latex}{\out{\hypertarget{method-bootstrap_once}{}}}
\subsection{Method \code{bootstrap_once()}}{
run one bootstrap sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityBootstrap$bootstrap_once(
  self,
  data,
  epsilon_step = self$epsilon_step,
  population_x = self$x,
  population_tmle = self$pointTMLE,
  inflate_lambda = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{self}}{self}

\item{\code{data}}{input bootstrap data}

\item{\code{epsilon_step}}{step size of the TMLE (one-step)}

\item{\code{population_x}}{whole training data}

\item{\code{population_tmle}}{TMLE on the whole training data}

\item{\code{inflate_lambda}}{experimental. scalar to scale the L1 penalty during the bootstrap}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-run_bootstrap}{}}}
\subsection{Method \code{run_bootstrap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityBootstrap$run_bootstrap(
  n_bootstrap = 200,
  alpha = 0.05,
  kind = NULL,
  inflate_lambda = 1,
  to_parallel = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityBootstrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
