% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgDensityValueTMLE.R
\docType{class}
\name{avgDensityTMLE}
\alias{avgDensityTMLE}
\title{onestep TMLE of average density parameter}
\format{
\code{\link{R6Class}} object.
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
onestep TMLE of average density parameter

onestep TMLE of average density parameter
}
\section{Methods}{

\describe{
  \item{\code{new(x, epsilon_step = NULL, verbose = NULL)}}{
     specify data; define targeting step size
  }
  \item{\code{fit_density(bin_width = .1, lambda_grid)}}{
     use `cvDensityHAL` to fit density. bin_width for pre-binning of
     continuous x; lambda_grid for grid search of lambda in HAL
  }
}
}

\examples{
# avgDensityTMLE$new()
}
\keyword{data}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{random sample from distribution}

\item{\code{p_hat}}{(\code{empiricalDensity}) containing density estimates}

\item{\code{Psi}}{paramter value}

\item{\code{EIC}}{vector of EIC}

\item{\code{epsilon_step}}{step size for one-step targeting}

\item{\code{CI}}{(numeric vector) length 2; lower + upper CI}

\item{\code{longDataOut}}{(\code{longiData}) for transforming x into longitudinal format dataframe}

\item{\code{HAL_tuned}}{(\code{hal9001}) generated by \code{cvDensityHAL} class.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{avgDensityTMLE$new()}}
\item \href{#method-fit_density}{\code{avgDensityTMLE$fit_density()}}
\item \href{#method-fit_density_pen_likeli}{\code{avgDensityTMLE$fit_density_pen_likeli()}}
\item \href{#method-compute_Psi}{\code{avgDensityTMLE$compute_Psi()}}
\item \href{#method-compute_EIC}{\code{avgDensityTMLE$compute_EIC()}}
\item \href{#method-update_once}{\code{avgDensityTMLE$update_once()}}
\item \href{#method-target_iterative}{\code{avgDensityTMLE$target_iterative()}}
\item \href{#method-target_onestep}{\code{avgDensityTMLE$target_onestep()}}
\item \href{#method-inference}{\code{avgDensityTMLE$inference()}}
\item \href{#method-compute_min_phi_ratio}{\code{avgDensityTMLE$compute_min_phi_ratio()}}
\item \href{#method-clone}{\code{avgDensityTMLE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$new(x, epsilon_step = NULL, verbose = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_density"></a>}}
\if{latex}{\out{\hypertarget{method-fit_density}{}}}
\subsection{Method \code{fit_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$fit_density(
  bin_width = 0.1,
  lambda_grid = NULL,
  M = NULL,
  n_fold = 3,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit_density_pen_likeli"></a>}}
\if{latex}{\out{\hypertarget{method-fit_density_pen_likeli}{}}}
\subsection{Method \code{fit_density_pen_likeli()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$fit_density_pen_likeli(
  bin_width = 0.1,
  lambda_grid = NULL,
  lambda_min_ratio = NULL,
  n_fold = 3,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_Psi"></a>}}
\if{latex}{\out{\hypertarget{method-compute_Psi}{}}}
\subsection{Method \code{compute_Psi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$compute_Psi(p_hat, to_return = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_EIC"></a>}}
\if{latex}{\out{\hypertarget{method-compute_EIC}{}}}
\subsection{Method \code{compute_EIC()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$compute_EIC(p_hat, Psi, to_return = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_once"></a>}}
\if{latex}{\out{\hypertarget{method-update_once}{}}}
\subsection{Method \code{update_once()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$update_once()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target_iterative"></a>}}
\if{latex}{\out{\hypertarget{method-target_iterative}{}}}
\subsection{Method \code{target_iterative()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$target_iterative(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target_onestep"></a>}}
\if{latex}{\out{\hypertarget{method-target_onestep}{}}}
\subsection{Method \code{target_onestep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$target_onestep(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inference"></a>}}
\if{latex}{\out{\hypertarget{method-inference}{}}}
\subsection{Method \code{inference()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$inference()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_min_phi_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-compute_min_phi_ratio}{}}}
\subsection{Method \code{compute_min_phi_ratio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$compute_min_phi_ratio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{avgDensityTMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
