% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_HAL.R
\name{basic_fixed_HAL}
\alias{basic_fixed_HAL}
\title{(Experimental) Super Learner wrapper for fixed HAL}
\usage{
basic_fixed_HAL(
  Y,
  X,
  hal9001_object = NULL,
  newX = NULL,
  family = stats::gaussian(),
  obsWeights = rep(1, length(Y)),
  inflate_lambda = 1,
  ...
)
}
\arguments{
\item{Y}{vector of target}

\item{X}{data.frame of feature}

\item{hal9001_object}{hal9001 object}

\item{newX}{data.frame of features to generate prediction}

\item{family}{glm family Y}

\item{obsWeights}{vector of weight}

\item{inflate_lambda}{scale the penalty factor}

\item{...}{other arguments}
}
\description{
depend on an hal9001 object. which is the fit on the whole data.
}
