% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blipVarianceTMLE.R
\name{blipVarianceTMLE}
\alias{blipVarianceTMLE}
\title{Compute TMLE on the variance of CATE (binary Y)}
\description{
Compute TMLE on the variance of CATE (binary Y)

Compute TMLE on the variance of CATE (binary Y)
}
\details{
inherit the initial fit routine from ate
}
\section{Super class}{
\code{\link[TMLEbootstrap:ateTMLE]{TMLEbootstrap::ateTMLE}} -> \code{blipVarianceTMLE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-initial_fit}{\code{blipVarianceTMLE$initial_fit()}}
\item \href{#method-target}{\code{blipVarianceTMLE$target()}}
\item \href{#method-compute_EIC}{\code{blipVarianceTMLE$compute_EIC()}}
\item \href{#method-inference_without_target}{\code{blipVarianceTMLE$inference_without_target()}}
\item \href{#method-compute_min_phi_ratio}{\code{blipVarianceTMLE$compute_min_phi_ratio()}}
\item \href{#method-clone}{\code{blipVarianceTMLE$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="initial_fit_pen_likeli">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-initial_fit_pen_likeli}{\code{TMLEbootstrap::ateTMLE$initial_fit_pen_likeli()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="initialize">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-initialize}{\code{TMLEbootstrap::ateTMLE$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="plot_Q1W">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-plot_Q1W}{\code{TMLEbootstrap::ateTMLE$plot_Q1W()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial_fit"></a>}}
\if{latex}{\out{\hypertarget{method-initial_fit}{}}}
\subsection{Method \code{initial_fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$initial_fit(family_y = "binomial", ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target"></a>}}
\if{latex}{\out{\hypertarget{method-target}{}}}
\subsection{Method \code{target()}}{
use the logistic submodel to target blip variance; output Psi, EIC, CI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$target()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_EIC"></a>}}
\if{latex}{\out{\hypertarget{method-compute_EIC}{}}}
\subsection{Method \code{compute_EIC()}}{
Compute EIC of blip variance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$compute_EIC(Y, A, Q1k, Q0k, Qk, gk, psi)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Y}}{continuous target variable}

\item{\code{A}}{binary treatment}

\item{\code{Q1k}}{E(Y|A=1,W)}

\item{\code{Q0k}}{E(Y|A=0,W)}

\item{\code{Qk}}{E(Y|A=a,W)}

\item{\code{gk}}{E(A=a|W)}

\item{\code{psi}}{var(E(Y|A=1,W) - E(Y|A=0,W))}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
EIC vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inference_without_target"></a>}}
\if{latex}{\out{\hypertarget{method-inference_without_target}{}}}
\subsection{Method \code{inference_without_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$inference_without_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_min_phi_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-compute_min_phi_ratio}{}}}
\subsection{Method \code{compute_min_phi_ratio()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$compute_min_phi_ratio()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
