% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blipVarianceTMLE.R
\name{blipVarianceTMLEContinuousY}
\alias{blipVarianceTMLEContinuousY}
\title{Compute TMLE on the variance of CATE (continuous Y)}
\description{
inherit the initial fit routine from ateTMLE
replace initial fit with continuous hal9001;
rescales Y into (0,1) before TMLE;
scales the Psi, EIC, CI back to the original scale after the TMLE
scale Y to (0,1)
}
\section{Super classes}{
\code{\link[TMLEbootstrap:ateTMLE]{TMLEbootstrap::ateTMLE}} -> \code{\link[TMLEbootstrap:blipVarianceTMLE]{TMLEbootstrap::blipVarianceTMLE}} -> \code{blipVarianceTMLEContinuousY}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-scale_outcome}{\code{blipVarianceTMLEContinuousY$scale_outcome()}}
\item \href{#method-scale_back_after_tmle}{\code{blipVarianceTMLEContinuousY$scale_back_after_tmle()}}
\item \href{#method-initial_fit}{\code{blipVarianceTMLEContinuousY$initial_fit()}}
\item \href{#method-target}{\code{blipVarianceTMLEContinuousY$target()}}
\item \href{#method-inference_without_target}{\code{blipVarianceTMLEContinuousY$inference_without_target()}}
\item \href{#method-clone}{\code{blipVarianceTMLEContinuousY$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="initial_fit_pen_likeli">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-initial_fit_pen_likeli}{\code{TMLEbootstrap::ateTMLE$initial_fit_pen_likeli()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="initialize">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-initialize}{\code{TMLEbootstrap::ateTMLE$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="ateTMLE" data-id="plot_Q1W">}\href{../../TMLEbootstrap/html/ateTMLE.html#method-plot_Q1W}{\code{TMLEbootstrap::ateTMLE$plot_Q1W()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="blipVarianceTMLE" data-id="compute_EIC">}\href{../../TMLEbootstrap/html/blipVarianceTMLE.html#method-compute_EIC}{\code{TMLEbootstrap::blipVarianceTMLE$compute_EIC()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="TMLEbootstrap" data-topic="blipVarianceTMLE" data-id="compute_min_phi_ratio">}\href{../../TMLEbootstrap/html/blipVarianceTMLE.html#method-compute_min_phi_ratio}{\code{TMLEbootstrap::blipVarianceTMLE$compute_min_phi_ratio()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale_outcome"></a>}}
\if{latex}{\out{\hypertarget{method-scale_outcome}{}}}
\subsection{Method \code{scale_outcome()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$scale_outcome()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale_back_after_tmle"></a>}}
\if{latex}{\out{\hypertarget{method-scale_back_after_tmle}{}}}
\subsection{Method \code{scale_back_after_tmle()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$scale_back_after_tmle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial_fit"></a>}}
\if{latex}{\out{\hypertarget{method-initial_fit}{}}}
\subsection{Method \code{initial_fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$initial_fit(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-target"></a>}}
\if{latex}{\out{\hypertarget{method-target}{}}}
\subsection{Method \code{target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-inference_without_target"></a>}}
\if{latex}{\out{\hypertarget{method-inference_without_target}{}}}
\subsection{Method \code{inference_without_target()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$inference_without_target()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{blipVarianceTMLEContinuousY$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
