% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comprehensiveBootstrap.R
\name{comprehensiveBootstrap}
\alias{comprehensiveBootstrap}
\title{Run `generalBootstrap` twice (regular + second-order bootstrap)}
\description{
Run `generalBootstrap` twice (regular + second-order bootstrap)

Run `generalBootstrap` twice (regular + second-order bootstrap)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{comprehensiveBootstrap$new()}}
\item \href{#method-bootstrap}{\code{comprehensiveBootstrap$bootstrap()}}
\item \href{#method-all_CI}{\code{comprehensiveBootstrap$all_CI()}}
\item \href{#method-compute_width}{\code{comprehensiveBootstrap$compute_width()}}
\item \href{#method-compute_wald_width}{\code{comprehensiveBootstrap$compute_wald_width()}}
\item \href{#method-clone}{\code{comprehensiveBootstrap$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$new(parameter = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bootstrap"></a>}}
\if{latex}{\out{\hypertarget{method-bootstrap}{}}}
\subsection{Method \code{bootstrap()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$bootstrap(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-all_CI"></a>}}
\if{latex}{\out{\hypertarget{method-all_CI}{}}}
\subsection{Method \code{all_CI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$all_CI()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_width"></a>}}
\if{latex}{\out{\hypertarget{method-compute_width}{}}}
\subsection{Method \code{compute_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$compute_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_wald_width"></a>}}
\if{latex}{\out{\hypertarget{method-compute_wald_width}{}}}
\subsection{Method \code{compute_wald_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$compute_wald_width()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{comprehensiveBootstrap$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
