% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityHAL.R
\name{cvDensityHAL}
\alias{cvDensityHAL}
\title{Fit a 1-d density using HAL regression; automatic tuning of L1 penalty}
\description{
perform binning first then run a HAL classifier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{cvDensityHAL$new()}}
\item \href{#method-assign_fold}{\code{cvDensityHAL$assign_fold()}}
\item \href{#method-cv}{\code{cvDensityHAL$cv()}}
\item \href{#method-cv_lambda_grid}{\code{cvDensityHAL$cv_lambda_grid()}}
\item \href{#method-compute_model_full_data}{\code{cvDensityHAL$compute_model_full_data()}}
\item \href{#method-clone}{\code{cvDensityHAL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$new(x, longiData)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign_fold"></a>}}
\if{latex}{\out{\hypertarget{method-assign_fold}{}}}
\subsection{Method \code{assign_fold()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$assign_fold(n_fold = 3)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv"></a>}}
\if{latex}{\out{\hypertarget{method-cv}{}}}
\subsection{Method \code{cv()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$cv(lambda = 2e-05, verbose = FALSE, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-cv_lambda_grid"></a>}}
\if{latex}{\out{\hypertarget{method-cv_lambda_grid}{}}}
\subsection{Method \code{cv_lambda_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$cv_lambda_grid(lambda_grid = NULL, lambda_min_ratio = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compute_model_full_data"></a>}}
\if{latex}{\out{\hypertarget{method-compute_model_full_data}{}}}
\subsection{Method \code{compute_model_full_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$compute_model_full_data(lambda)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{cvDensityHAL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
