% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityHAL.R
\name{densityHAL}
\alias{densityHAL}
\title{Fit a 1-d density using HAL regression}
\description{
perform binning first then run a HAL classifier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{densityHAL$new()}}
\item \href{#method-fit}{\code{densityHAL$fit()}}
\item \href{#method-predict}{\code{densityHAL$predict()}}
\item \href{#method-predict_long}{\code{densityHAL$predict_long()}}
\item \href{#method-eval_misclass_loss}{\code{densityHAL$eval_misclass_loss()}}
\item \href{#method-eval_crossentropy_loss}{\code{densityHAL$eval_crossentropy_loss()}}
\item \href{#method-clone}{\code{densityHAL$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$new(x, longiData)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$fit(lambda = 2e-05, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$predict(new_x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_long"></a>}}
\if{latex}{\out{\hypertarget{method-predict_long}{}}}
\subsection{Method \code{predict_long()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$predict_long(new_x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_misclass_loss"></a>}}
\if{latex}{\out{\hypertarget{method-eval_misclass_loss}{}}}
\subsection{Method \code{eval_misclass_loss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$eval_misclass_loss(new_x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_crossentropy_loss"></a>}}
\if{latex}{\out{\hypertarget{method-eval_crossentropy_loss}{}}}
\subsection{Method \code{eval_crossentropy_loss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$eval_crossentropy_loss(new_x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{densityHAL$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
