% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empiricalDensity}
\alias{empiricalDensity}
\title{Store a 1-dimensional density function}
\description{
Store a 1-dimensional density function

Store a 1-dimensional density function
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{empiricalDensity$new()}}
\item \href{#method-normalize}{\code{empiricalDensity$normalize()}}
\item \href{#method-display}{\code{empiricalDensity$display()}}
\item \href{#method-clone}{\code{empiricalDensity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empiricalDensity$new(p_density, x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-normalize"></a>}}
\if{latex}{\out{\hypertarget{method-normalize}{}}}
\subsection{Method \code{normalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empiricalDensity$normalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display"></a>}}
\if{latex}{\out{\hypertarget{method-display}{}}}
\subsection{Method \code{display()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empiricalDensity$display(p_truth = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{empiricalDensity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
