% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_HAL.R
\name{fit_fixed_HAL}
\alias{fit_fixed_HAL}
\title{fitting fixed_HAL. outputs an object of the fit
use the old basis
use the old lambda
OPTIONAL: if the old object is squashed, only use the non-zero basis}
\usage{
fit_fixed_HAL(
  Y,
  X,
  weights = NULL,
  hal9001_object,
  family = stats::gaussian(),
  inflate_lambda = 1
)
}
\arguments{
\item{Y}{vector of target}

\item{X}{data.frame of feature}

\item{weights}{vector of weight}

\item{hal9001_object}{hal9001 object}

\item{family}{glm family Y}

\item{inflate_lambda}{scale the penalty factor}
}
\description{
fitting fixed_HAL. outputs an object of the fit
use the old basis
use the old lambda
OPTIONAL: if the old object is squashed, only use the non-zero basis
}
