% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_HAL.R
\name{generate_SL.fixed_HAL}
\alias{generate_SL.fixed_HAL}
\title{Generator of SL wrappers}
\usage{
generate_SL.fixed_HAL(hal9001_object = NULL, inflate_lambda = 1)
}
\description{
outputs a SL wrapper, that no longer depend on the hal9001 object.
the output arguments conform with `SL` library convention
}
\keyword{internal}
