% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneHyperparam.R
\name{grabPlateau}
\alias{grabPlateau}
\title{Grab a plateau of a function y = f(x)}
\description{
Grab a plateau of a function y = f(x)

Grab a plateau of a function y = f(x)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{grabPlateau$new()}}
\item \href{#method-find_plateau}{\code{grabPlateau$find_plateau()}}
\item \href{#method-clone}{\code{grabPlateau$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grabPlateau$new(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-find_plateau"></a>}}
\if{latex}{\out{\hypertarget{method-find_plateau}{}}}
\subsection{Method \code{find_plateau()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grabPlateau$find_plateau(find_plateau_start = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{grabPlateau$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
