% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longiData}
\alias{longiData}
\title{Convert univariate series to longitudinal format}
\description{
helper for avgDens TMLE.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{longiData$new()}}
\item \href{#method-generate_df}{\code{longiData$generate_df()}}
\item \href{#method-generate_df_compress}{\code{longiData$generate_df_compress()}}
\item \href{#method-clone}{\code{longiData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{longiData$new(x, bin_width = 0.5)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_df"></a>}}
\if{latex}{\out{\hypertarget{method-generate_df}{}}}
\subsection{Method \code{generate_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{longiData$generate_df(x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_df_compress"></a>}}
\if{latex}{\out{\hypertarget{method-generate_df_compress}{}}}
\subsection{Method \code{generate_df_compress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{longiData$generate_df_compress(x = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{longiData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
