% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_HAL.R
\name{predict.fixed_HAL}
\alias{predict.fixed_HAL}
\title{prediciton function for fixed_HAL object}
\usage{
\method{predict}{fixed_HAL}(object, ..., new_data)
}
\arguments{
\item{object}{class `fixed_HAL``}

\item{...}{extra arguments into hal9001}

\item{new_data}{matrix with the same number of columns as in training}
}
\value{
a vector of predictions
}
\description{
prediciton function for fixed_HAL object
}
\keyword{internal}
