% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_HAL.R
\name{predict.SL.fixed_HAL}
\alias{predict.SL.fixed_HAL}
\title{SuperLearner prediction function for the SL wrapper}
\usage{
\method{predict}{SL.fixed_HAL}(object, newX, ...)
}
\description{
SuperLearner prediction function for the SL wrapper
}
\keyword{internal}
