% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaleX}
\alias{scaleX}
\title{Perform min/max standardization based on (x - min)/(max - min)}
\description{
helper for blipVarianceTMLEContinuousY
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{scaleX$new()}}
\item \href{#method-scale01}{\code{scaleX$scale01()}}
\item \href{#method-scaleOriginal}{\code{scaleX$scaleOriginal()}}
\item \href{#method-clone}{\code{scaleX$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaleX$new(X)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scale01"></a>}}
\if{latex}{\out{\hypertarget{method-scale01}{}}}
\subsection{Method \code{scale01()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaleX$scale01(newX = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-scaleOriginal"></a>}}
\if{latex}{\out{\hypertarget{method-scaleOriginal}{}}}
\subsection{Method \code{scaleOriginal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaleX$scaleOriginal(newX = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{scaleX$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
