% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneHyperparam.R
\name{tuneHyperparam}
\alias{tuneHyperparam}
\title{Use plateau method to choose the L1 penalty of HAL (abstract class)}
\description{
Use plateau method to choose the L1 penalty of HAL (abstract class)

Use plateau method to choose the L1 penalty of HAL (abstract class)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{tuneHyperparam$new()}}
\item \href{#method-get_lambda}{\code{tuneHyperparam$get_lambda()}}
\item \href{#method-get_value}{\code{tuneHyperparam$get_value()}}
\item \href{#method-plot_CI}{\code{tuneHyperparam$plot_CI()}}
\item \href{#method-get_lambda_df}{\code{tuneHyperparam$get_lambda_df()}}
\item \href{#method-plot_width}{\code{tuneHyperparam$plot_width()}}
\item \href{#method-select_lambda_pleateau_wald}{\code{tuneHyperparam$select_lambda_pleateau_wald()}}
\item \href{#method-get_Psi_df}{\code{tuneHyperparam$get_Psi_df()}}
\item \href{#method-select_lambda_psi_grad}{\code{tuneHyperparam$select_lambda_psi_grad()}}
\item \href{#method-clone}{\code{tuneHyperparam$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$new(data, n_bootstrap = 200, inflate_lambda = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-get_lambda}{}}}
\subsection{Method \code{get_lambda()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$get_lambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_value"></a>}}
\if{latex}{\out{\hypertarget{method-get_value}{}}}
\subsection{Method \code{get_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$get_value()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_CI"></a>}}
\if{latex}{\out{\hypertarget{method-plot_CI}{}}}
\subsection{Method \code{plot_CI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$plot_CI(Psi = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_lambda_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_lambda_df}{}}}
\subsection{Method \code{get_lambda_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$get_lambda_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_width"></a>}}
\if{latex}{\out{\hypertarget{method-plot_width}{}}}
\subsection{Method \code{plot_width()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$plot_width(type_CI = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select_lambda_pleateau_wald"></a>}}
\if{latex}{\out{\hypertarget{method-select_lambda_pleateau_wald}{}}}
\subsection{Method \code{select_lambda_pleateau_wald()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$select_lambda_pleateau_wald(df_lambda_width)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_Psi_df"></a>}}
\if{latex}{\out{\hypertarget{method-get_Psi_df}{}}}
\subsection{Method \code{get_Psi_df()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$get_Psi_df()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-select_lambda_psi_grad"></a>}}
\if{latex}{\out{\hypertarget{method-select_lambda_psi_grad}{}}}
\subsection{Method \code{select_lambda_psi_grad()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$select_lambda_psi_grad(df_Psi)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{tuneHyperparam$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
