% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageDensityValueTMLE.R
\docType{class}
\name{avgDensityTMLE}
\alias{avgDensityTMLE}
\title{onestep TMLE of average density parameter}
\format{\code{\link{R6Class}} object.}
\usage{
avgDensityTMLE
}
\value{
Object of \code{\link{R6Class}} with methods
}
\description{
onestep TMLE of average density parameter
}
\section{Fields}{

\describe{
\item{\code{x}}{random sample from distribution}

\item{\code{p_hat}}{(\code{empiricalDensity}) containing density estimates}

\item{\code{Psi}}{paramter value}

\item{\code{EIC}}{vector of EIC}

\item{\code{epsilon_step}}{step size for one-step targeting}

\item{\code{CI}}{(numeric vector) length 2; lower + upper CI}

\item{\code{longDataOut}}{(\code{longiData}) for transforming x into longitudinal format dataframe}

\item{\code{HAL_tuned}}{(\code{hal9001}) generated by \code{cv_densityHAL} class.}
}}

\section{Methods}{

\describe{
  \item{\code{new(x, epsilon_step = NULL, verbose = NULL)}}{specify data; define targeting step size}
  \item{\code{fit_density(bin_width = .1, lambda_grid)}}{use `cv_densityHAL` to fit density. bin_width for pre-binning of continuous x; lambda_grid for grid search of lambda in HAL}
}
}

\examples{
# avgDensityTMLE$new()
}
\keyword{data}
